Create Steam_iFrame.html
This is an iFrame for Steam. I tried doing everything all in one page, but I couldn;t due to Sandbox issues and whatnot. So this way is the best way to display of Steam which then this will be an iFrame on gabegaming.com. Hope it works since it took me 3 months and lots of Google/StackOGreFlow.